New U-Net dataGenerator: simpler and more powerful
